+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # MMM-voice Changelog
 
+## [1.0.2]
+
+### Fixed
+
+* Restored context for handleData, logDebug and logError
+* DeprecationWarning: Calling async function without callback (noop function)
+
 ## [1.0.1]
 
 ### Added